move some tests to scan_teste
